requirements.txt
certifi==2019.9.11
chardet==3.0.4
Click==7.0
cycler==0.10.0
Flask==1.1.1
gunicorn==20.0.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.0
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
newsapi-python==0.2.5
numpy==1.17.3
pyparsing==2.4.4
python-dateutil==2.8.1
requests==2.21.0
scikit-learn==0.21.3
scipy==1.3.2
six==1.13.0
sklearn==0.0
urllib3==1.24.3
Werkzeug==0.16.0
